requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
bidict==0.23.1
build==1.2.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
docutils==0.21.2
fastapi==0.115.5
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
importlib-metadata==8.5.0
jaraco-classes==3.4.0
jaraco-context==6.0.1
jaraco-functools==4.1.0
jinja2==3.1.4
keyring==25.5.0
lazy-loader==0.4
mako==1.3.6
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
more-itertools==10.5.0
nh3==0.2.18
packaging==24.2
pip==24.3.1
pipdeptree==2.16.2
pkginfo==1.10.0
platformdirs==4.3.6
psutil==6.1.0
psycopg2-binary==2.9.10
pydantic==2.10.1
pydantic-core==2.27.1
pygments==2.18.0
pyproject-hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-multipart==0.0.17
python-socketio==5.11.4
readme-renderer==44.0
redis==5.2.0
reflex==0.6.5
reflex-ag-grid==0.0.10
reflex-chakra==0.6.2
reflex-hosting-cli==0.1.18
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
setuptools==75.6.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
sqlalchemy==2.0.36
sqlmodel==0.0.22
starlette==0.41.3
starlette-admin==0.14.1
tabulate==0.9.0
tomlkit==0.13.2
twine==5.1.1
typer==0.13.1
typing-extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
websockets==14.1
wheel==0.45.0
wrapt==1.17.0
wsproto==1.2.0
zipp==3.21.0
